+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookCategoryController.java
@@ -3,7 +3,7 @@ package com.nhnacademy.frontend.admin.controller;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryCreateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryUpdateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookCategoryResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
