+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookSearchController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.frontend.admin.controller;
 
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookSearchResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import groovy.util.logging.Slf4j;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
