+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookTagMapController.java
@@ -2,7 +2,7 @@ package com.nhnacademy.frontend.admin.controller;
 
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookTagMapCreateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookTagMapResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
