+++ src/main/java/com/nhnacademy/frontend/book/controller/BookController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.frontend.book.controller;
 
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookDetailResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
