+++ src/test/java/com/nhnacademy/frontend/book/controller/BookCategoryControllerTest.java
@@ -4,7 +4,7 @@ import com.nhnacademy.frontend.admin.controller.BookCategoryController;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryCreateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryUpdateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookCategoryResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import com.nhnacademy.frontend.auth.filter.JwtAuthenticationFilter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
