+++ src/test/java/com/nhnacademy/frontend/book/controller/BookTagControllerTest.java
@@ -3,7 +3,7 @@ package com.nhnacademy.frontend.book.controller;
 import com.nhnacademy.frontend.admin.controller.BookTagController;
 import com.nhnacademy.frontend.common.adapter.dto.book.request.BookTagCreateRequestDto;
 import com.nhnacademy.frontend.common.adapter.dto.book.response.BookTagResponseDto;
-import com.nhnacademy.frontend.admin.service.BookService;
+import com.nhnacademy.frontend.common.service.BookService;
 import com.nhnacademy.frontend.auth.filter.JwtAuthenticationFilter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
